# docs/swagger.yaml
definitions:
...
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    format: int32
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
...
  reqmodel.FavoriteCreateRequest:
    properties:
      movie_id:
        type: string
    type: object
  reqmodel.FavoriteDeleteRequest:
    properties:
      movie_id:
        type: string
    type: object
  reqmodel.FavoriteGetRequest:
    properties:
      movie_id:
        type: string
      user_id:
        type: string
    type: object
...
  reqmodel.MovieCreateRequest:
    properties:
      title:
        type: string
    type: object
  reqmodel.MovieFavoriteListResponse:
    properties:
      favorite_movie_ids:
        items:
          type: string
        type: array
      movie_id:
        type: string
    type: object
  reqmodel.MovieListResponse:
    properties:
      movie_list:
        items:
          $ref: '#/definitions/sqlc.Movie'
        type: array
    type: object
...
  reqmodel.MovieUpdateRequest:
    properties:
      title:
        type: string
    type: object
...
  reqmodel.UserFavoriteListResponse:
    properties:
      favorite_user_ids:
        items:
          type: string
        type: array
      user_id:
        type: string
    type: object
...
  sqlc.Favorite:
    properties:
      movie_id:
        type: string
      user_id:
        type: string
    type: object
...
  sqlc.Movie:
    properties:
      created_at:
        $ref: '#/definitions/pgtype.Timestamp'
      id:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Basic swagger for current api
  termsOfService: http://swagger.io/terms/
  title: movie_backend_go
  version: "1.0"
paths:
...
  /favorite:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Comment delete data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.FavoriteDeleteRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Delete favorite
      tags:
      - favorite
    get:
      consumes:
      - application/json
      parameters:
      - description: Favorite data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.FavoriteGetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.Favorite'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get favorite
      tags:
      - favorite
    post:
      consumes:
      - application/json
      description: Add movie to user's favorite
      parameters:
      - description: Comment update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.FavoriteCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.Favorite'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Add favorite movie
      tags:
      - favorite
...
  /movie:
    get:
      consumes:
      - application/json
      description: Get all movie list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqmodel.MovieListResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get movie list
      tags:
      - movie
    post:
      consumes:
      - application/json
      parameters:
      - description: Movie creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.MovieCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sqlc.Movie'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Create movie
      tags:
      - movie
      - admin
  /movie/{movie_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Delete movie
      tags:
      - movie
      - admin
    get:
      consumes:
      - application/json
      description: Get movie by id
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.Movie'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get movie
      tags:
      - movie
    patch:
      consumes:
      - application/json
      description: Update movie
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: string
      - description: Movie creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reqmodel.MovieUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sqlc.Movie'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Update movie
      tags:
      - movie
      - admin
...
  /movie/{movie_id}/favorite:
    get:
      consumes:
      - application/json
      description: Get list users who marked this movie as favorite
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqmodel.MovieFavoriteListResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get movie favorite list
      tags:
      - favorite
      - movie
...
  /user/{user_id}/favorite:
    get:
      consumes:
      - application/json
      description: Get user's favorite movie list
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reqmodel.UserFavoriteListResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user favorite list
      tags:
      - favorite
      - user
...
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: http://localhost:8080/auth/login
    type: oauth2
swagger: "2.0"